Split source hydration apply phase
