+++ src/app/controller/library/sources/hydration.rs
@@ -3,14 +3,13 @@ use super::hydration_telemetry;
 #[cfg(not(test))]
 use crate::app::controller::jobs::JobMessage;
 use crate::app::controller::jobs::{
-    SourceHydrationJob, SourceHydrationKind, SourceHydrationResult, SourceHydrationSnapshot,
+    SourceHydrationJob, SourceHydrationKind, SourceHydrationResult,
 };
-use crate::app::controller::state::cache::FolderBrowserCacheKey;
 use crate::app::controller::state::runtime::PendingSourceHydration;
 use crate::app::state::{FolderBrowserUiState, FolderPaneId};
-use std::path::PathBuf;
 use std::time::Instant;
 
+mod apply;
 mod worker;
 
 use worker::{run_source_hydration, source_hydration_async_enabled};
@@ -120,206 +119,6 @@ impl AppController {
             }
         }
         hydration_telemetry::record_source_hydration_apply(apply_start.elapsed());
-    }
-
-    fn apply_source_hydration_snapshot(
-        &mut self,
-        kind: SourceHydrationKind,
-        pane: FolderPaneId,
-        source_id: SourceId,
-        elapsed: std::time::Duration,
-        snapshot: SourceHydrationSnapshot,
-    ) {
-        let from_cache = snapshot.from_cache;
-        let deferred_follow_up_work = snapshot.deferred_follow_up_work;
-        match kind {
-            SourceHydrationKind::ActiveSelection => {
-                if !self.install_hydrated_wav_entries(&source_id, &snapshot) {
-                    self.finish_source_loading(kind, pane);
-                    return;
-                }
-                if let Some(feature_cache) = snapshot.feature_cache.clone() {
-                    self.install_feature_cache_snapshot(source_id.clone(), feature_cache);
-                } else {
-                    self.queue_feature_cache_refresh_for_browser();
-                }
-                if deferred_follow_up_work {
-                    self.queue_folder_browser_refresh();
-                } else {
-                    let available_folders = snapshot.available_folders;
-                    let folder_tree = snapshot.folder_tree;
-                    self.apply_folder_snapshot_to_pane(
-                        pane,
-                        &source_id,
-                        &self
-                            .current_source()
-                            .expect("selected source should exist")
-                            .root,
-                        available_folders,
-                        folder_tree,
-                    );
-                }
-                self.apply_post_source_hydration_selection();
-                self.finish_source_hydration_metadata(&source_id, from_cache, elapsed);
-                if self.should_rebuild_browser_lists_async() {
-                    self.dispatch_search_job();
-                    if let Some(pending) =
-                        self.runtime.source_lane.hydration.pending_active.as_mut()
-                    {
-                        pending.search_request_id =
-                            Some(self.ui.browser.search.latest_search_request_id);
-                    }
-                } else {
-                    self.rebuild_browser_lists();
-                    self.finish_source_loading(kind, pane);
-                }
-            }
-            SourceHydrationKind::InactivePane => {
-                let available_folders = snapshot.available_folders;
-                let folder_tree = snapshot.folder_tree;
-                let Some(source) = self
-                    .library
-                    .sources
-                    .iter()
-                    .find(|source| source.id == source_id)
-                    .cloned()
-                else {
-                    self.finish_source_loading(kind, pane);
-                    return;
-                };
-                self.apply_folder_snapshot_to_pane(
-                    pane,
-                    &source_id,
-                    &source.root,
-                    available_folders,
-                    folder_tree,
-                );
-                self.finish_source_loading(kind, pane);
-            }
-        }
-    }
-
-    fn apply_folder_snapshot_to_pane(
-        &mut self,
-        pane: FolderPaneId,
-        source_id: &SourceId,
-        source_root: &std::path::Path,
-        available: std::collections::BTreeSet<PathBuf>,
-        tree: crate::app::controller::library::source_folders::FolderTreeSnapshot,
-    ) {
-        let key = FolderBrowserCacheKey {
-            pane,
-            source_id: source_id.clone(),
-        };
-        let model = {
-            let model = self.ui_cache.folders.models.entry(key.clone()).or_default();
-            model.reconcile_available(source_root, available);
-            model.clone()
-        };
-        self.ui_cache.folders.snapshots.insert(key, tree.clone());
-        self.queue_folder_projection_with_snapshot(pane, source_id.clone(), model, tree);
-    }
-
-    fn install_hydrated_wav_entries(
-        &mut self,
-        source_id: &SourceId,
-        snapshot: &SourceHydrationSnapshot,
-    ) -> bool {
-        self.wav_entries.total = snapshot.total;
-        self.wav_entries.page_size = snapshot.page_size.max(1);
-        self.wav_entries.pages.clear();
-        self.wav_entries.lookup = snapshot.path_lookup.clone();
-        self.wav_entries.source_id = Some(source_id.clone());
-        self.wav_entries.pages.insert(0, snapshot.entries.clone());
-        if self
-            .wav_entries
-            .pages
-            .get(&0)
-            .is_some_and(|page| page.iter().any(|entry| entry.missing))
-            && let Some(source) = self
-                .library
-                .sources
-                .iter()
-                .find(|source| &source.id == source_id)
-                .cloned()
-        {
-            let missing_paths = self
-                .wav_entries
-                .pages
-                .get(&0)
-                .into_iter()
-                .flat_map(|page| page.iter())
-                .filter(|entry| entry.missing)
-                .map(|entry| entry.relative_path.clone())
-                .collect::<Vec<_>>();
-            for path in missing_paths {
-                let _ = self.prune_missing_sample(&source, &path);
-            }
-            return false;
-        }
-        self.ui_cache.browser.search.invalidate();
-        self.ui_cache.browser.pipeline.invalidate();
-        true
-    }
-
-    fn apply_post_source_hydration_selection(&mut self) {
-        let mut pending_applied = false;
-        if let Some(path) = self.runtime.jobs.take_pending_select_path() {
-            if self.sample_view.wav.selected_wav.as_ref() == Some(&path) {
-                pending_applied = true;
-            } else if self.wav_index_for_path(&path).is_some() {
-                self.select_wav_by_path_with_rebuild(&path, false);
-                pending_applied = true;
-            }
-        }
-        if !pending_applied
-            && self.sample_view.wav.selected_wav.is_none()
-            && self.wav_entries.total > 0
-        {
-            self.selection_state.suppress_autoplay_once = true;
-            self.select_wav_by_index_with_rebuild(0, false);
-        }
-    }
-
-    fn finish_source_hydration_metadata(
-        &mut self,
-        source_id: &SourceId,
-        from_cache: bool,
-        elapsed: std::time::Duration,
-    ) {
-        if !from_cache {
-            self.ui_cache.browser.labels.remove(source_id);
-            self.ui_cache.browser.bpm_values.remove(source_id);
-        }
-        let needs_failures = !from_cache
-            || !self
-                .ui_cache
-                .browser
-                .analysis_failures
-                .contains_key(source_id);
-        if needs_failures {
-            if let Some(source) = self
-                .library
-                .sources
-                .iter()
-                .find(|source| &source.id == source_id)
-                .cloned()
-            {
-                self.queue_analysis_failures_refresh(&source);
-            } else {
-                self.ui_cache.browser.analysis_failures.remove(source_id);
-            }
-        }
-        self.sync_missing_from_db(source_id);
-        self.set_background_status(
-            format!(
-                "Loaded {} wav files in {} ms",
-                self.wav_entries.total,
-                elapsed.as_millis()
-            ),
-            StatusTone::Info,
-        );
-        crate::app::controller::library::wavs::apply_pending_similarity_filter_rebuild(self);
     }
 
     pub(crate) fn clear_active_source_for_loading(&mut self) {
